+++ dsr_agents/CHANGELOG.rst
@@ -16,6 +16,7 @@ X.X.X (XX-02-2024)
 * Remove QtExecutor as it's integrated in dsr_util.
 * Move ros_to_dsr_types.hpp to dsr_util.
 * Move agent_node.hpp to dsr_util and move agents.
+* Skip the removal of person if they are interacting.
 
 0.6.2 (23-01-2024)
 ------------------